+++ Src/Log/DlgFilter.cpp
@@ -1,3 +1,6 @@
+// Copyright Copyright (c) Kang Lin studio, All Rights Reserved
+// Author Kang Lin <kl222@126.com>
+
 #include "DlgFilter.h"
 #include "ui_DlgFilter.h"
 #include <QDebug>