+++ ChangeLog.md
@@ -4,7 +4,7 @@
   - Add CDlgEdit
   - Add CFileBrowser
   - Add file browser application
-  - Add print stack trace
+  - Add stack trace
 
 ### Version v2.2.6
   - Modify debian package files